# backend/docs/swagger.yaml
basePath: /
definitions:
...
  CreateGoldRes:
    properties:
      data:
        $ref: '#/definitions/domain.Gold'
      profit:
        format: double
        type: number
      sell_price:
        format: double
        type: number
    required:
    - data
    - profit
    - sell_price
    type: object
  CreateGoldTaxPreferenceRes:
    properties:
      preference:
        $ref: '#/definitions/dto.GoldTaxPreference'
    required:
    - preference
    type: object
...
  DeleteGoldRes:
    properties:
      message:
        type: string
    required:
    - message
    type: object
  DeleteGoldTaxPreferenceRes:
    properties:
      message:
        type: string
    required:
    - message
    type: object
...
  GetGoldRes:
    properties:
      data:
        $ref: '#/definitions/domain.Gold'
      profit:
        format: double
        type: number
      sell_price:
        format: double
        type: number
    required:
    - data
    - profit
    - sell_price
    type: object
...
  GoldDashboardRes:
    properties:
      data:
        $ref: '#/definitions/dto.GoldDashboard'
    required:
    - data
    type: object
  ListGoldTaxPreferencesRes:
    properties:
      preferences:
        items:
          $ref: '#/definitions/dto.GoldTaxPreference'
        type: array
    required:
    - preferences
    type: object
...
  UpdateGoldRes:
    properties:
      data:
        $ref: '#/definitions/domain.Gold'
      profit:
        format: double
        type: number
      sell_price:
        format: double
        type: number
    required:
    - data
    - profit
    - sell_price
    type: object
  UpdateGoldTaxPreferenceRes:
    properties:
      preference:
        $ref: '#/definitions/dto.GoldTaxPreference'
    required:
    - preference
    type: object
...
  domain.FxRate:
    properties:
      base:
        type: string
      created_at:
        type: string
      date:
        type: string
      id:
        type: integer
      rate:
        type: integer
    required:
    - base
    - created_at
    - date
    - id
    - rate
    type: object
  domain.Gold:
    properties:
      carat:
        format: double
        type: number
      created_at:
        type: string
      date:
        type: string
      grams:
        format: double
        type: number
      id:
        type: integer
      notes:
        type: string
      price:
        description: price per gram in IDR
        format: double
        type: number
      type:
        description: pure_gold or jewelry
        type: string
      updated_at:
        type: string
      wallet_id:
        type: integer
    required:
    - carat
    - created_at
    - date
    - grams
    - id
    - notes
    - price
    - type
    - updated_at
    - wallet_id
    type: object
  domain.GoldPrice:
    properties:
      created_at:
        type: string
      currency:
        type: string
      currency_symbol:
        type: string
      exchange_rate:
        format: double
        type: number
      id:
        type: integer
      price_per_ounce_usd:
        format: double
        type: number
      source_readable_age:
        type: string
      source_updated_at:
        type: string
      symbol:
        type: string
    required:
    - created_at
    - currency
    - currency_symbol
    - exchange_rate
    - id
    - price_per_ounce_usd
    - source_readable_age
    - source_updated_at
    - symbol
    type: object
  domain.GoldTaxPreference:
    properties:
      carat:
        format: double
        type: number
      created_at:
        type: string
      id:
        type: integer
      tax_rate:
        format: double
        type: number
      updated_at:
        type: string
      user_id:
        type: integer
    required:
    - carat
    - created_at
    - id
    - tax_rate
    - updated_at
    - user_id
    type: object
...
  dto.Error:
    properties:
      code:
        type: integer
      message:
        type: string
    required:
    - code
    - message
    type: object
  dto.Gold:
    properties:
      data:
        $ref: '#/definitions/domain.Gold'
      profit:
        format: double
        type: number
      sell_price:
        format: double
        type: number
    required:
    - data
    - profit
    - sell_price
    type: object
  dto.GoldDashboard:
    properties:
      fx_rate:
        $ref: '#/definitions/domain.FxRate'
      gold_price:
        $ref: '#/definitions/domain.GoldPrice'
      gold_taxes:
        items:
          $ref: '#/definitions/dto.GoldTax'
        type: array
      latest_golds:
        items:
          $ref: '#/definitions/dto.Gold'
        type: array
      profit:
        format: double
        type: number
      retail_price:
        format: double
        type: number
      total_asset:
        format: double
        type: number
      total_buy_price:
        format: double
        type: number
      total_gold_items:
        type: integer
      total_weight:
        format: double
        type: number
    required:
    - fx_rate
    - gold_price
    - gold_taxes
    - latest_golds
    - profit
    - retail_price
    - total_asset
    - total_buy_price
    - total_gold_items
    - total_weight
    type: object
  dto.GoldTax:
    properties:
      data:
        $ref: '#/definitions/domain.GoldTaxPreference'
      sell_price:
        format: double
        type: number
    required:
    - data
    - sell_price
    type: object
  dto.GoldTaxPreference:
    properties:
      carat:
        format: double
        type: number
      created_at:
        type: string
      id:
        type: integer
      tax_rate:
        format: double
        type: number
      updated_at:
        type: string
      user_id:
        type: integer
    required:
    - carat
    - created_at
    - id
    - tax_rate
    - updated_at
    - user_id
    type: object
...
  gold.CreateGoldReq:
    properties:
      carat:
        format: double
        type: number
      date:
        type: string
      grams:
        format: double
        type: number
      notes:
        type: string
      price:
        format: double
        type: number
      type:
        type: string
      wallet_id:
        type: integer
    required:
    - carat
    - date
    - grams
    - notes
    - price
    - type
    - wallet_id
    type: object
  gold.GoldTaxPreferenceReq:
    properties:
      carat:
        format: double
        type: number
      tax_rate:
        format: double
        type: number
    required:
    - carat
    - tax_rate
    type: object
  gold.ListGoldsRes:
    properties:
      golds:
        items:
          $ref: '#/definitions/dto.Gold'
        type: array
    required:
    - golds
    type: object
  gold.UpdateGoldReq:
    properties:
      carat:
        format: double
        type: number
      date:
        type: string
      grams:
        format: double
        type: number
      notes:
        type: string
      price:
        format: double
        type: number
      type:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API untuk aplikasi manajemen keuangan dan pencatatan emas - ARTA
  title: ARTA API
  version: "1.0"
paths:
...
  /api/dashboard/gold:
    get:
      consumes:
      - application/json
      operationId: GetGoldDashboard
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: wallet_id
        in: query
        name: wallet_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/GoldDashboardRes'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.Error'
      tags:
      - dashboard
...
  /api/gold:
    get:
      consumes:
      - application/json
      operationId: ListGolds
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gold.ListGoldsRes'
      tags:
      - gold
    post:
      consumes:
      - application/json
      operationId: CreateGold
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Unique key per submission attempt for safe retry (UUID recommended)
        in: header
        name: Idempotency-Key
        type: string
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/gold.CreateGoldReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/CreateGoldRes'
      tags:
      - gold
  /api/gold/{id}:
    delete:
      consumes:
      - application/json
      operationId: DeleteGold
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: gold id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/DeleteGoldRes'
      tags:
      - gold
    get:
      consumes:
      - application/json
      operationId: GetGold
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: gold id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/GetGoldRes'
      tags:
      - gold
    put:
      consumes:
      - application/json
      operationId: UpdateGold
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: gold id
        in: path
        name: id
        required: true
        type: integer
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/gold.UpdateGoldReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/UpdateGoldRes'
      tags:
      - gold
  /api/gold/tax:
    get:
      consumes:
      - application/json
      operationId: ListGoldTaxPreferences
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ListGoldTaxPreferencesRes'
      tags:
      - gold
    post:
      consumes:
      - application/json
      operationId: CreateGoldTaxPreference
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Unique key per submission attempt for safe retry (UUID recommended)
        in: header
        name: Idempotency-Key
        type: string
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/gold.GoldTaxPreferenceReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/CreateGoldTaxPreferenceRes'
      tags:
      - gold
  /api/gold/tax/{id}:
    delete:
      consumes:
      - application/json
      operationId: DeleteGoldTaxPreference
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: tax preference id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/DeleteGoldTaxPreferenceRes'
      tags:
      - gold
    put:
      consumes:
      - application/json
      operationId: UpdateGoldTaxPreference
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: tax preference id
        in: path
        name: id
        required: true
        type: integer
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/gold.GoldTaxPreferenceReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/UpdateGoldTaxPreferenceRes'
      tags:
      - gold
...
schemes:
- http
- https
securityDefinitions:
  Bearer:
    description: '"Type ''Bearer TOKEN''"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"